fix details row and cleanup redundant types
