+++ src/components/CreateCampaign/StepFour/StepFour.tsx
@@ -31,14 +31,6 @@ const StepFour = () => {
       { title: 'Creatives', value: adUnitsFormatted, isColumn: true },
       { title: 'Selected Categories', value: formattedCats },
       { title: 'Selected Countries', value: formattedLocs },
-      // {
-      //   title: 'Include incentivized traffic',
-      //   value: advancedTargeInput.includeIncentivized ? 'Yes' : 'No'
-      // },
-      // {
-      //   title: 'Disable frequency capping',
-      //   value: advancedTargeInput.disableFrequencyCapping ? 'Yes' : 'No'
-      // },
       {
         title: 'Limit average daily spending',
         value: advancedTargeInput.limitDailyAverageSpending ? 'Yes' : 'No'
