+++ src/types/index.ts
@@ -15,5 +15,4 @@ export * from './selectItem'
 export * from './topUpWithFiat'
 export * from './timeFrameChart'
 export * from './usePagination'
-export * from './campaignDetails'
 export * from './campaignsData'
